最新branch merge後ユーザ退会仕様作成完了

+++ app/Http/Controllers/UsersController.php
@@ -4,22 +4,30 @@ namespace App\Http\Controllers;
 
 use App\User;
 use Illuminate\Support\Facades\Auth;
-
+use Illuminate\Http\Request;
 class UsersController extends Controller
 {
     public function show($id)
     {
         $user = User::findOrFail($id);
-        $posts = $user->posts()->orderBy('id','desc')->paginate(6);
+        $posts = $user->posts()->orderBy('id', 'desc')->paginate(6);
 
-        return view('users.detail',compact('user','posts'));
+        return view('users.detail', compact('user', 'posts'));
     }
 
-    public function destroy($id)
+    public function destroy($id, Request $request)
     {
         $user = User::findOrFail($id);
+        if(Auth::id() !== $user->id){
+            abort(403,'権限がありません。');
+        }
         $user->delete();
 
-        return redirect('/');
+        Auth::logout();
+
+        $request->session()->invalidate(); // session無効化
+        $request->session()->regenerateToken(); // CSRFトークンをリセット
+
+        return redirect('/'); // 退会後トップページへ
     }
 }
